feat: macOS agent service — native providers, remote commands, scheme handler
- Add CoreFoundationNative interop for native macOS API access
- Update MacOS infrastructure providers: ActiveWindow, FilePathExtractor,
  IdleDetector, MachineMetrics, Notifications, KeychainTokenStore
- Add MacOSRemoteCommandExecutor for remote command handling
- Update TimeTrackSchemeHandler with full API proxy implementation
- Upgrade project files to net10.0 target framework
- Fix gitignore: exclude src/build/ (compiled app bundle artifacts)

+++ src/agent/TimeTrack.Agent.Infrastructure.MacOS/Providers/MacOSIdleDetector.cs
@@ -1,4 +1,5 @@
 using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using TimeTrack.Agent.Contracts.Providers;
@@ -79,8 +80,8 @@ public sealed class MacOSIdleDetector : IIdleDetector
         try
         {
             var idleSeconds = CGEventSourceSecondsSinceLastEventType(
-                CGEventSourceStateID.CombinedSessionState,
-                CGEventType.AllEvents);
+                (int)CGEventSourceStateID.CombinedSessionState,
+                (int)CGEventType.AllEvents);
 
             if (idleSeconds < 0)
             {
